send tourist income from tourist to ui

+++ PutYourHandsUpInTheAir/Assets/Systems/GameMessages/Messages/ReducePotentialIncomeAction.cs
@@ -2,6 +2,6 @@ namespace Systems.GameMessages.Messages
 {
     public class ReducePotentialIncomeAction
     {
-        public int IncomeVanished { get; set; }
+        public float IncomeVanished { get; set; }
     }
 }
